Move the interpolation functions and friends to interpolaters.py
It's only DRY in the case of topaz4_interpolate; the other functions
are only used in era5_topaz4_maker.py, but this way I've seperated file
handling and interpolation and rotation.
